feat: adiciona implementações

+++ vocachat/src/main/java/com/github/gerdanyjr/vocachat/application/dto/in/assessment/CreateAssessmentRequest.java
@@ -1,4 +1,4 @@
 package com.github.gerdanyjr.vocachat.application.dto.in.assessment;
 
-public record CreateAssessmentRequest(Integer userId) {
+public record CreateAssessmentRequest(Long userId) {
 }
