+++ vocachat/src/main/java/com/github/gerdanyjr/vocachat/application/repository/user/IUserRepository.java
@@ -5,5 +5,5 @@ import com.github.gerdanyjr.vocachat.core.model.User;
 import java.util.Optional;
 
 public interface IUserRepository {
-    Optional<User> findById(Integer userId);
+    Optional<User> findById(Long userId);
 }
